Fix for passing ssh private key to argument
I have a script that gets run like this:

```
node upload-key < path/to/key
```

I then internally convert that blob to an argument for commander, which ends up being like this:

```
process.argv
// [ 'node', 'upload-key', '--key', '-----BEGIN RSA PRIVATE KEY-----\nMx1....' ]
```

It seems that just adding a check to see if there are more than two `--` dashes will do the trick (it works for me locally). Thoughts?