add callback to allow filtering or post-processing of help text
added a callback to allow auto-generated help text to be filtered or
post-processed.  for example, i want to apply colors to the text before
it is displayed in the console.

add callback to allow filtering or post-processing of help text

fix brace alignment in example

make brace alignment in example consistent with other code

add callback to allow filtering or post-processing of help text

added a callback to allow auto-generated help text to be filtered or
post-processed.  for example, i want to apply colors to the text before
it is displayed in the console.

add callback to allow filtering or post-processing of help text

fix brace alignment in example

make brace alignment in example consistent with other code