Fix Coercion for Non Collections
This fixes a bug that appears to have been introduced when
repeated options were added that causes the previous value or
default value to be passed as a second parameter to the coercion
function. This obviously causes unintended side effects when a user
specifies a function and a default value unaware that a second value
passed, such as with parseInt.

This change makes the assumption that when a default array is specified
that the user has expecting repeated options. This change is backward
compatible with the existing code.