Emit new events beforeAction and beforeExec before internal and (respectively) external command executions.
The command tree is a hierarchy, and a "parent" command may wnat to intercede
or report on some or all of its subcommands. For this reason, the emitted
event names add additional prefixes at each level. My use-case application,
schematool, has a command hierarchy

schematool
  |
  +-- typescript-types
  |
  +-- ....

At the typescript-types node, the emitted events are

   beforeAction:*
   beforeAction:typescript-types

At the schematool node, the emitted events are
   beforeAction:*
   beforeAction:schematool.*
   beforeAction:schematool.typescript-types