Simplified the _emitToTree logic per my note to John Gee. No more expanding list of event actions.
Renamed beforeAction to preAction and beforeExec to preExec.
Added postAction

Looked at adding postExec, but it isn't entirely clear where this should
fire. I would naively expect it to fire after the subordinate process
completes, but best to check with John first, as there's no point
sticking it in the wrong place and moving it later.

I note that the proc object is getting preserved below the
proc.close/proc.error, but the saved field is not referenced anywhere.
Is that dead code?