Return proxy from constructor further up the prototype chain
Borrowed from eb142d8 that was supposed to land in the now-closed #1921.

The change ensures the this object used in Command is the proxy from the
very beginning, even in the constructor. This solves pretty much all
issues with returning the proxy from a constructor. For example, private
fields can be used now. More details available at:
https://github.com/tj/commander.js/pull/1921#issuecomment-1661035391

Additionally, the ownKeys() trap and wrong spelling in comments have
been fixed (the former change being borrowed from fc927c8).