Remove test of default export in TypeScript
In the test file, an aggregating import (import * as commander) was used
to allegedly import the default export. That is not the same as actually
importing the default export (import commander).

Fixing the import revealed the following error:

TS1192: Module '".../commander.js/typings/index"' has no default export.

The reason is that the default export is not publicly declared because
it is deprecated, and so should not be used.

The test was therefore unnecessary.